requirements.txt
azure-mgmt-compute
azure-identity
pandas
openpyxl
azure-mgmt-monitor
msrestazure
azure-mgmt-advisor==9.0.0b1
azure-mgmt-resource==15.0.0
azure-loganalytics
azure-core
